backend/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
Flask==2.2.5
flask-cors==3.0.10


